refactor for agents
