--- evidence_project/.evidence/customization/.profile.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:4c016a5d678b9f134d65408fcd44b5779682dd51aee4ec82d647586b17a53c63
-size 111
