--- evidence_project/.evidence/customization/custom-formatting.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:185765d0ccabcfdcc56e0644aa5f08dc3fdf648834d822bb39fd7b73a687990a
-size 48
