--- evidence_project/.gitignore
@@ -1,9 +1,0 @@
-.evidence/template
-.svelte-kit
-build
-node_modules
-.DS_Store
-static/data
-*.options.yaml
-.vscode/settings.json
-.env
