--- evidence_project/.npmrc
@@ -1,3 +1,0 @@
-loglevel=error
-audit=false
-fund=false
