--- evidence_project/.vscode/extensions.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:84d9936f036ec0a48845179f7201c1d2d026f049c53e5fcd335c41522e04c4ee
-size 61
