--- evidence_project/evidence.plugins.yaml
@@ -1,42 +1,0 @@
-
-components:
-  # This loads all of evidence's core charts and UI components
-  # You probably don't want to edit this dependency unless you know what you are doing
-  "@evidence-dev/core-components": {}
-
-  # Loading a plugin:
-  # your-component-plugin: {}
-  #
-  # Adding aliases:
-  # your-component-plugin:
-  #   aliases:
-  #     SomeLongComponent: Alias
-  #
-  # Adding overrides:
-  # your-component-plugin:
-  #   aliases:
-  #     SomeLongChartComponent: BarChart
-  #   overrides:
-  #     - BarChart
-  #
-  # Adding normal Svelte component library:
-  # your-svelte-component-library:
-  #   provides:
-  #     - SomeComponent
-  #     - SomeOtherComponent
-  #
-  # For more info: https://docs.evidence.dev/plugins/using-plugins/
-datasources:
-  # You can add additional datasources here by adding npm packages. 
-  # Make to also add them to `package.json`.
-  "@evidence-dev/bigquery": { }
-  "@evidence-dev/csv": { }
-  "@evidence-dev/databricks": { }
-  "@evidence-dev/duckdb": { }
-  "@evidence-dev/mssql": { }
-  "@evidence-dev/mysql": { }
-  "@evidence-dev/postgres": { }
-  "@evidence-dev/snowflake": { }
-  "@evidence-dev/sqlite": { }
-  "@evidence-dev/trino": { }
-  "@evidence-dev/motherduck": { }
