--- evidence_project/geo_json/county.geo.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:61ca9938f12a68dbbee040f7565a0b91db3d74e457e11626a5c270c63c8deb6f
-size 174838130
