--- evidence_project/geo_json/us-states.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:6f23ed91fce2c25d57c01d83225342ba40258e9571c486c3999bd827c80d193b
-size 89263
