--- evidence_project/package.json
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f3d5c886cd5dfd73b4b2673605574c16ff4019a7991dce30bb007e6aaeae2eb9
-size 1107
