--- evidence_project/queries/govt_economic_stats/fred_series_latest_aggregates.sql
@@ -1,3 +1,0 @@
-SELECT
-    *
-FROM econ_md.prod_econ.fred_series_latest_aggregates
