--- evidence_project/queries/govt_economic_stats/housing_inventory.sql
@@ -1,3 +1,0 @@
-SELECT
-    *
-FROM econ_md.prod_econ.housing_inventory
