--- evidence_project/queries/maps/country_map_data.sql
@@ -1,3 +1,0 @@
-SELECT
-    *
-FROM econ_md.prod_econ.country_map_data
