--- evidence_project/queries/maps/county_map_data.sql
@@ -1,3 +1,0 @@
-SELECT
-    *
-FROM "econ_md"."prod_econ"."stg_realtor_county_history"
