--- evidence_project/queries/maps/state_map_data.sql
@@ -1,3 +1,0 @@
-select
-    * 
-from econ_md.prod_econ.latest_state_data
