--- evidence_project/sources/econ/connection.yaml
@@ -1,4 +1,0 @@
-name: econ_db
-type: duckdb
-options:
-  filename: econ_db.duckdb
