--- evidence_project/sources/econ_md/connection.yaml
@@ -1,5 +1,0 @@
-# This file was automatically generated
-name: econ_md
-type: motherduck
-options:
-  database: prod_econ
