--- evidence_project/sources/econ_md/raw/fred_data.sql
@@ -1,3 +1,0 @@
-SELECT
-    *
-FROM main.stg_fred_series
