--- evidence_project/sources/econ_md/raw/housing_inventory.sql
@@ -1,3 +1,0 @@
-SELECT 
-    *
-FROM main.stg_housing_inventory
