--- evidence_project/sources/econ_md/raw/housing_pulse.sql
@@ -1,3 +1,0 @@
-SELECT
-    *
-FROM main.stg_housing_pulse
