--- evidence_project/sources/econ_md/raw/state_data.sql
@@ -1,3 +1,0 @@
-SELECT
-  *
-FROM main.stg_realtor_state_history
