+++ macro_agents/src/macro_agents/defs/assets/ingestion/schwab.py
@@ -1,0 +1,84 @@
+## ToDo: Implement Schwab ingestion assets
+
+major_indicies_tickers = [
+    "SPY",
+    "QQQ",
+    "DIA",
+    "IWM",
+]
+## Need to find a way to get the vix
+
+## US Sector etfs
+us_sector_etfs = [
+    "XLK",
+    "XLC",
+    "XLY",
+    "XLF",
+    "XLI",
+    "XLU",
+    "XLP",
+    "XLRE",
+    "XLB",
+    "XLE",
+    "XLV",
+]
+## Fixed income etfs
+fixed_income_etfs = [
+    "CWB",
+    "HYG",
+    "LQD",
+    "TIP",
+    "GOVT",
+    "MUB",
+]
+
+## Pull in currency crosses
+# BTC pound, yen, euro
+
+
+## find the etfs for these if not use the futures
+energy_commodities = [
+    "NG1",
+    "DL1",
+    "HO1",
+    "XB1",
+    "CL1",
+    "CO1",
+]
+
+# metals find the etfs for these if not use the futures
+metal_commodities = [
+    "silver",
+    "gold",
+    "copper",
+    "steel",
+    "aluminum",
+]
+
+# agriculture
+agriculture_commodities = [
+    "corn",
+    "wheat",
+    "soybeans",
+    "feeder cattle",
+    "live cattle",
+    "sugar",
+    "cotton",
+]
+
+# global markets
+global_markets = [
+    "EEM",
+    "ACWI",
+    "EWG",
+    "EWJ",
+    "EWU",
+    "EWQ",
+    "EWA",
+    "FXI",
+    "EZA",
+    "EWY",
+    "EWW",
+    "EWZ",
+    "EPI",
+]
