+++ macro_agents/src/macro_agents/defs/resources.py
@@ -1,0 +1,5 @@
+from .resources.motherduck import motherduck_resource
+from .resources.fred import fred_resource
+
+md = motherduck_resource
+fred = fred_resource
