+++ macro_agents/src/macro_agents/defs/resources/fred.py
