+++ macro_agents/src/macro_agents/defs/resources/schwab.py
@@ -1,0 +1,1 @@
+## ToDo: Implement Schwab resource
